+++ package.json
@@ -4,10 +4,11 @@
   "description": "Manage users and admins list",
   "main": "index.js",
   "scripts": {
-    "dev": "npx tsc && node ./build/index.js",
+    "dev": "npm-run-all --parallel dev:*",
+    "dev-tsc": "npx tsc && node ./build/index.js",
     "dev:server": "nodemon --watch build --exec \"node build/bundle.js\"",
-    "dev:build:server": "webpack --config webpack.server.js --watch",
-    "dev:build:client": "webpack --config webpack.client.js --watch"
+    "dev:build-server": "webpack --config webpack.server.js --watch",
+    "dev:build-client": "webpack --config webpack.client.js --watch"
   },
   "author": "",
   "license": "ISC",
