+++ webpack.server.js
@@ -1,6 +1,9 @@
 const path = require('path')
+const merge = require('webpack-merge')
+const baseConfig = require('./webpack.base')
+const webpackNodeExternals = require('webpack-node-externals')
 
-module.exports = {
+const config = {
   // Target env on where to run the output code
   target: 'node',
 
@@ -12,22 +15,7 @@ module.exports = {
     filename: 'bundle.js',
     path: path.resolve(__dirname, 'build')
   },
-
-  // babel setup - need to be same on both server and client webpack
-  module: {
-    rules: [
-      {
-        test: /\.js?$/,
-        loader: 'babel-loader',
-        exclude: '/node_modules/',
-        options: {
-          presets: [
-            'react',
-            'stage-0',
-            ['env', { targets: { browsers: ['last 2 versions'] } }]
-          ]
-        }
-      }
-    ]
-  }
+  externals: [webpackNodeExternals()]
 }
+
+module.exports = merge(baseConfig, config)
